+++ src/config/test.js
@@ -1,13 +1,21 @@
+/**
+ * https://www.npmjs.com/package/config
+ * $ npm install config
+ */
 const host = process.env.DB_HOST || 'localhost';
+const dbname = 'node-copybanking-test';
 
 module.exports = {
   server: {
-    port: 9001
+    port: 27017
   },
   database: {
-    url: `mongodb://${host}/node-express-skeleton-test`,
+    url: `mongodb://${host}/${dbname}`,
     properties: {
-      useMongoClient: true
+      //useMongoClient: true,
+      useNewUrlParser: true,
+      useUnifiedTopology: true,
+      useCreateIndex: true
     }
   },
   key: {
