+++ src/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "com.accenture.sec",
+  "version": "1.0.0",
+  "description": "Node + Express REST API skeleton application",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "ssh://git@devops01.secreloaded.sec:7999/fnd/fwk-archetype-nodejs.git"
+  },
+  "author": "SEC",
+  "license": "ISC",
+  "dependencies": {
+    "bcrypt": "^3.0.2",
+    "body-parser": "^1.19.0",
+    "config": "^3.2.4",
+    "express": "^4.17.1",
+    "jsonwebtoken": "^8.5.1",
+    "mongoose": "^5.8.7",
+    "mongoose-paginate": "^5.0.3",
+    "morgan": "^1.9.1",
+    "pow-mongodb-fixtures": "^0.14.0"
+  }
+}
